src/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
neo4j==5.12.0
geoserver-rest==2.5.1
python-dotenv==1.0.0
neo4j-driver==5.12.0
